+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [#36](https://github.com/itk-dev/devops_itksites/pull/36)
+  Implemented OIDC code flow and handled target path after login.
+
 ## [1.4.0] - 2023-08-01
 
 - [#34](https://github.com/itk-dev/devops_itksites/pull/34)